Initialize subgraph
