+++ subgraphs/chamber/package.json
@@ -1,17 +1,13 @@
 {
   "name": "chamber",
-  "description": "Chamber Subgraph",
-  "version": "1.0.0",
-  "repository": "git@github.com:looksrare/subgraphs.git",
-  "author": "Loreum",
-  "license": "MIT",
+  "license": "UNLICENSED",
   "scripts": {
     "codegen": "graph codegen",
     "build": "graph build",
-    "deploy": "graph deploy --node https://api.studio.thegraph.com/deploy/ registry-chamber",
-    "create-local": "graph create --node http://localhost:8020/ registry-chamber",
-    "remove-local": "graph remove --node http://localhost:8020/ registry-chamber",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 registry-chamber",
+    "deploy": "graph deploy --node https://api.studio.thegraph.com/deploy/ chamber",
+    "create-local": "graph create --node http://localhost:8020/ chamber",
+    "remove-local": "graph remove --node http://localhost:8020/ chamber",
+    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 chamber",
     "test": "graph test"
   },
   "dependencies": {
